server/requirements-frozen.txt
amqp==5.0.2
aniso8601==8.1.0
asgiref==3.3.1
auth0-python==3.14.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
blinker==1.4
Brotli==1.0.9
bs4==0.0.1
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.2
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
dnspython==1.16.0
edlib==1.3.8.post2
extras==1.0.0
Faker==5.4.1
filetype==1.0.7
fixtures==3.0.0
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Cors==3.0.10
Flask-HTTPAuth==4.2.0
Flask-JWT-Extended==3.25.0
Flask-RESTful==0.3.8
giturlparse==0.10.0
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-core==1.5.0
google-cloud-logging==2.0.2
google-cloud-storage==1.35.0
google-cloud-secret-manager==2.2.0
google-crc32c==1.1.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.35.0rc1
gunicorn==20.0.4
h11==0.12.0
h2==4.0.0
hpack==4.0.0
httplib2==0.18.1
hyperframe==6.0.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
jwcrypto==0.8
kaitaistruct==0.9
kiwisolver==1.3.1
kombu==5.0.2
ldap3==2.8.1
linecache2==1.0.0
MarkupSafe==1.1.1
matplotlib==3.3.3
mitmproxy==5.3.0
mongoengine==0.22.1
msgpack==1.0.2
networkx==2.5
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.5.1.48
pandas==1.1.5
passlib==1.7.4
pbr==5.5.1
Pillow==8.1.0
prompt-toolkit==3.0.10
proto-plus==1.13.0
protobuf==3.13.0
psutil==5.8.0
publicsuffix2==2.20191221
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.1
python-dateutil==2.8.1
python-http-client==3.3.1
python-mimeparse==1.6.0
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.3.1
questionary==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-image==0.17.2
scikit-learn==0.24.0
scipy==1.5.4
selenium==4.0.0a6.post2
sendgrid==6.4.8
six==1.15.0
slack-webhook==1.0.4
sortedcontainers==2.2.2
soupsieve==2.1
starkbank-ecdsa==1.1.0
stripe==2.55.1
testtools==2.4.0
text-unidecode==1.3
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.7.1
tornado==6.1
traceback2==1.4.0
typing-extensions==3.7.4.3
unittest2==1.1.0
uritemplate==3.0.1
urllib3==1.26.2
urwid==2.1.2
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
wsproto==0.15.0
zipp==3.4.0
zstandard==0.14.1
